+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Development
   - e.g. TEMPERATURE_AIR_MEAN_2M, CLOUD_COVER_BETWEEN_2KM_TO_7KM, PROBABILITY_PRECIPITATION_HEIGHT_GT_0_0MM_LAST_6H
 - Bump pyarrow version to <18
 - Fix EaHydrology station list parsing
+- Rename `EaHydrology` to `EAHydrology`
 
 0.91.0 (14.07.2024)
 *******************